requirements.txt
black==21.11b1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cssselect==1.1.0
cssutils==2.3.0
cycler==0.11.0
fonttools==4.28.1
idna==3.3
kiwisolver==1.3.2
lxml==4.6.4
matplotlib==3.5.0
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.21.4
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
premailer==3.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
regex==2021.11.10
requests==2.26.0
setuptools-scm==6.3.2
six==1.16.0
tomli==1.2.2
tqdm==4.62.3
typing_extensions==4.0.0
urllib3==1.26.7
yagmail==0.14.260
